requirements.txt
pandas==1.2.4
numpy==1.19.5
tensorflow==2.4.1
tensorflow_probability
gin-config==0.5.0
pyaml
gtfparse==1.2.1
tensorflow_probability==0.12.2
papermill
nbconvert
pysam==0.18.0
pyBigWig==0.3.16
creme==0.6.1
protobuf==3.20.0
igrads @ git+https://github.com/mhorlacher/igrads.git@0.3.0
biopython